# backend/docs/swagger.yaml
definitions:
...
  Role:
    properties:
      casbinRole:
        description: Casbin-Role
        example: admin
        type: string
      createdAt:
        description: 创建时间
        example: "2006-01-02 15:04:05"
        type: string
      id:
        description: ID
        type: integer
      name:
        description: 角色名
        example: Admin
        type: string
      updatedAt:
        description: 更新时间
        example: "2006-01-02 15:04:05"
        type: string
    type: object
...
  User:
    properties:
      avatarUrl:
        description: 头像
        example: https://example.com/avatar.jpg
        type: string
      bio:
        description: 简介
        example: The Jackal
        type: string
      createdAt:
        description: 创建时间
        example: "2006-01-02 15:04:05"
        type: string
      direction:
        description: 方向
        example: ltr
        type: string
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
      fullName:
        description: 全名
        example: Zhang San
        type: string
      language:
        description: 语言
        example: zh-CN
        type: string
      phone:
        description: 手机
        example: "13800138000"
        type: string
      roles:
        description: 角色
        items:
          $ref: '#/definitions/Role'
        type: array
      status:
        description: 状态 0:待激活 1:正常 2:禁用
        example: 1
        type: integer
      theme:
        description: 主题
        example: light
        type: string
      timezone:
        description: 时区
        example: Asia/Shanghai
        type: string
      updatedAt:
        description: 更新时间
        example: "2006-01-02 15:04:05"
        type: string
      userId:
        description: UserID
        example: IQrIkwYlyn
        type: string
      username:
        description: 用户名
        example: zhangsan
        type: string
    type: object
  UserList:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/User'
        type: array
      total:
        description: 总数
        type: integer
    type: object
...
  backend_api_v1.DynamicMenuResponse:
    properties:
      data:
        $ref: '#/definitions/backend_api_v1.DynamicMenuResponseData'
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
  backend_api_v1.DynamicMenuResponseData:
    properties:
      list:
        description: 顶级菜单
        items:
          $ref: '#/definitions/backend_api_v1.MenuNode'
        type: array
    type: object
...
  backend_api_v1.MenuNode:
    properties:
      access:
        description: 可见性
        example: canAdmin
        type: string
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/backend_api_v1.MenuNode'
        type: array
      component:
        description: 组件
        example: '@/pages/Admin/User'
        type: string
      createdAt:
        description: 创建时间
        example: "2006-01-02 15:04:05"
        type: string
      disabled:
        type: boolean
      disabledTooltip:
        type: boolean
      flatMenu:
        description: 隐藏自身+子节点提升并打平
        type: boolean
      hideChildrenInMenu:
        description: 隐藏子节点
        type: boolean
      hideInMenu:
        description: 隐藏自身和子节点
        type: boolean
      icon:
        description: 图标
        example: crown
        type: string
      id:
        description: ID
        example: 1
        type: integer
      key:
        type: string
      locale:
        description: 国际化
        type: string
      name:
        description: 名称
        example: User
        type: string
      parentId:
        description: 父级菜单
        example: 0
        type: integer
      parentKeys:
        type: string
      path:
        description: 路径
        example: /admin/user
        type: string
      redirect:
        description: 重定向
        type: string
      target:
        description: 指定外链打开形式
        type: string
      tooltip:
        type: string
      updatedAt:
        description: 更新时间
        example: "2006-01-02 15:04:05"
        type: string
    type: object
...
  backend_api_v1.Response:
    properties:
      data:
        description: 返回数据
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
...
  backend_api_v1.UpdatePasswordRequest:
    properties:
      newPassword:
        description: 新密码
        example: "123456"
        type: string
      oldPassword:
        description: 旧密码
        example: "123456"
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  backend_api_v1.UpdateStatusRequest:
    properties:
      status:
        description: 状态 0:待激活 1:正常 2:禁用
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
    required:
    - status
    type: object
  backend_api_v1.UserRequest:
    properties:
      bio:
        description: 简介
        example: The Jackal
        type: string
      direction:
        description: 方向
        example: ltr
        type: string
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
      fullName:
        description: 全名
        example: Zhang San
        type: string
      language:
        description: 语言
        example: zh-CN
        type: string
      phone:
        description: 手机
        example: "13800138000"
        type: string
      roles:
        description: 角色
        items:
          type: string
        type: array
      status:
        description: 状态 0:待激活 1:正常 2:禁用
        example: 1
        type: integer
      theme:
        description: 主题
        example: light
        type: string
      timezone:
        description: 时区
        example: Asia/Shanghai
        type: string
      username:
        description: 用户名
        example: zhangsan
        type: string
    type: object
  backend_api_v1.UserResponse:
    properties:
      data:
        $ref: '#/definitions/User'
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
  backend_api_v1.UserSearchResponse:
    properties:
      data:
        $ref: '#/definitions/UserList'
      errorCode:
        description: 错误码
        type: integer
      errorMessage:
        description: 报错信息
        type: string
      errorShowType:
        description: 前端展示方式
        type: integer
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
host: localhost:7001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vibe Testing
  version: 1.0.0
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: 搜索时支持用户名、昵称、手机和邮箱筛选
      operationId: ListUsers
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 分页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 手机
        in: query
        name: phone
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 展示名
        in: query
        name: fullName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.UserSearchResponse'
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - User
    post:
      consumes:
      - application/json
      description: 创建一个新的用户
      operationId: CreateUser
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - User
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的用户
      operationId: DeleteUser
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 更新用户信息
      operationId: UpdateUser
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - User
  /admin/users/{id}/reset-avatar:
    put:
      consumes:
      - application/json
      description: 重置用户头像为默认头像
      operationId: ResetAvatar
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 重置头像
      tags:
      - User
  /admin/users/{id}/revoke-sessions:
    post:
      consumes:
      - application/json
      description: 撤销指定用户的所有登录会话
      operationId: RevokeSessions
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 撤销登录态
      tags:
      - User
  /admin/users/{id}/send-reset-email:
    post:
      consumes:
      - application/json
      description: 向指定用户发送重置密码邮件
      operationId: SendResetEmail
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 发送重置密码邮件
      tags:
      - User
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新用户状态（启用/禁用）
      operationId: UpdateStatus
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 更新用户状态
      tags:
      - User
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定ID的用户详情
      operationId: GetUserByID
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.UserResponse'
      security:
      - Bearer: []
      summary: 获取用户详情
      tags:
      - User
  /users/menu:
    get:
      consumes:
      - application/json
      description: 获取当前用户的菜单列表
      operationId: FetchDynamicMenu
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.DynamicMenuResponse'
      security:
      - Bearer: []
      summary: 获取用户菜单
      tags:
      - User
  /users/password:
    put:
      consumes:
      - application/json
      description: 更新用户密码
      operationId: UpdatePassword
      parameters:
      - description: 更新密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 更新密码
      tags:
      - User
  /users/profile:
    get:
      consumes:
      - application/json
      description: 获取当前用户的详细信息
      operationId: FetchCurrentUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.UserResponse'
      security:
      - Bearer: []
      summary: 获取当前用户
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 更新用户信息
      operationId: UpdateProfile
      parameters:
      - description: 参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_api_v1.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - User
  /users/profile/avatar:
    post:
      consumes:
      - application/json
      description: 上传用户头像
      operationId: UploadAvatar
      parameters:
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_api_v1.Response'
      security:
      - Bearer: []
      summary: 上传头像
      tags:
      - User
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"